@@ -1,12 +1,9 @@
-﻿using System.Collections.Generic;
+﻿namespace SharedApi.Services;
 
-namespace SharedApi.Services
+public class SomeSharedService
 {
-    public class SomeSharedService
+    public List<string> GetData()
     {
-        public List<string> GetData()
-        {
-            return new List<string> { "data from shared api", "more data" };
-        }
+        return new List<string> { "data from shared api", "more data" };
     }
 }
